pyproject.toml
[project]
name = "iaptoolkit"
version = "0.3.7-a4"
description = "Library of common utils for interacting with Identity-Aware Proxies"
authors = [
    {name = "Rob Voigt", email = "code@ravoigt.com"}
]
readme = "README.md"
requires-python = ">=3.11"

# [tool.poetry]
# packages = [{include = "iaptoolkit"}]

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://github.com/RAVoigt/iaptoolkit"
Repository = "https://github.com/RAVoigt/iaptoolkit"

# ================================
# Tools etc.
[tool.black]
line-length = 110
# target-version = ['py311'] # Weirdly unsupported currently
include = '\.pyi?$'

# ================================
# Dependencies
[tool.poetry.dependencies]
python = "^3.11"
google-auth = "^2.29.0"
requests = "^2.31.0"
toml = "^0.10.2"
kvcommon = {extras = ["k8s"], version = "^0.2.12"}

# flask = { version = "^0.20.0", optional = true }
# quart = { version = "^0.20.0", optional = true }
# prometheus-client = { version = "^0.20.0", optional = true }

[tool.poetry.group.dev.dependencies]
black = "*"
ipdb = "*"
ipython = "*"
mock = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
pytest-socket = "*"
pyfakefs = "^5.3.2"

flask = "*"
quart = "*"
prometheus_client = "*"
